Move merge logic to SemanticFieldMapper.Builder
